Bug10927: bootstrap: keep user's config file
Prevent removing Reviewboard config file on bootstrap

It amends changeset 1d1f47b28b

+++ bootstrap.sh
@@ -64,9 +64,9 @@ for dname, _, _ in os.walk('.'):
             pass
 EOF
 elif test -d .git; then
-    # Remove files ignored by Git
+    # Remove files ignored by Git, excluding user's config files
     for i in $(git ls-files --others --ignored --exclude-standard) ; do
-        rm --force $i
+        test "$i" = ".reviewboardrc" || rm --force $i
     done
 else
     echo "Unknown type of repository"
